move netcdf parser test to it's own module
